+++ src/onegov.org/onegov/org/assets/js/tags-input.js
@@ -1,3 +1,4 @@
+/* via https://github.com/developit/tags-input/pull/19 */
 (function(global, factory) {
     if (typeof define === 'function' && define.amd) {
         define(['exports', 'module'], factory);
@@ -87,13 +88,6 @@
             if (el)
                 el.classList.add('selected');
         }
-        function setInputWidth() {
-            var last = $('.tag', true).pop(),
-                w = base.offsetWidth;
-            if (!w)
-                return;
-            base.input.style.width = Math.max(w - (last ? last.offsetLeft + last.offsetWidth : 5) - 5, w / 4) + 'px';
-        }
         function savePartialInput(value) {
             if (typeof value !== 'string' && !Array.isArray(value)) {
                 value = base.input.value;
@@ -101,7 +95,6 @@
             if (addTag(value) !== false) {
                 base.input.value = '';
                 save();
-                setInputWidth();
             }
         }
         function refocus(e) {
@@ -147,7 +140,6 @@
                 selectedTag = $('.tag.selected'),
                 pos = el.selectionStart === el.selectionEnd && el.selectionStart,
                 last = $('.tag', true).pop();
-            setInputWidth();
             if (key === ENTER || key === COMMA || key === TAB) {
                 if (!el.value && key !== COMMA)
                     return;
@@ -156,13 +148,11 @@
                 if (selectedTag.nextSibling !== base.input)
                     select(selectedTag.nextSibling);
                 base.removeChild(selectedTag);
-                setInputWidth();
                 save();
             } else if (key === BACKSPACE) {
                 if (selectedTag) {
                     select(selectedTag.previousSibling);
                     base.removeChild(selectedTag);
-                    setInputWidth();
                     save();
                 } else if (last && pos === 0) {
                     select(last);